requirements.txt
numpy
pandas
click
Pillow
scikit-image
torch
torchvision